requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
fvcore
thop
tqdm
jupyterlab
